requirements.txt
Django==1.5.1
django-model-utils==1.2.0
djangorestframework==2.2.7
requests==1.2.0
wsgiref==0.1.2
